% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decidePar.R
\name{decidePar}
\alias{decidePar}
\title{Calibrate CBHM variance parameters (a, b)}
\usage{
decidePar(cohortsize, ntype, ntrial, p0, p1, var.small, var.big)
}
\arguments{
\item{cohortsize}{Matrix of interim cohort sizes (rows = looks, cols = baskets).}

\item{ntype}{Number of baskets.}

\item{ntrial}{Number of simulated trials for calibration.}

\item{p0}{Vector of null response rates by basket.}

\item{p1}{Vector of alternative response rates by basket.}

\item{var.small}{Pre-defined "small" variance target.}

\item{var.big}{Pre-defined "large" variance target.}
}
\value{
A list with elements \code{a} and \code{b}.
}
\description{
Computes calibration parameters for the calibrated Bayesian hierarchical model (CBHM)
by linking the between-basket variance to a homogeneity statistic \eqn{T} using:
\deqn{\sigma^2 = \exp(a + b \log(T)).}
}
